Extracted drawSoundOnCanvas into a helper file, ready to be re-used.
